Чиним добавление продукта на дату отличную от сегодня, показываем диалог "Вы уверены?" при таком добавлении (https://trello.com/c/GeKjTDey)
